+++ package.json
@@ -83,7 +83,7 @@
     "@rollup/plugin-node-resolve": "^13.3.0",
     "@rollup/plugin-replace": "^4.0.0",
     "@rollup/plugin-url": "^7.0.0",
-    "@tdesign/test": "workspace:^",
+    "@tdesign/internal-tests": "workspace:^",
     "@testing-library/dom": "^8.20.1",
     "@types/babel__traverse": "~7.20.6",
     "@types/lodash-es": "^4.17.12",
