+++ packages/components/tag-input/__tests__/vitest-tag-input.test.jsx
@@ -8,7 +8,7 @@ import { mount } from '@vue/test-utils';
 import { vi } from 'vitest';
 import { TagInput } from '..';
 import { getTagInputValueMount, getTagInputDefaultMount } from './mount';
-import { simulateInputChange, simulateInputEnter } from '@tdesign/test/utils';
+import { simulateInputChange, simulateInputEnter } from '@tdesign/internal-tests/utils';
 
 describe('TagInput Component', () => {
   it('props.clearable: empty TagInput does not need clearIcon', async () => {
