+++ packages/components/upload/__tests__/vitest-upload.test.jsx
@@ -7,7 +7,7 @@
 import { mount } from '@vue/test-utils';
 import { vi } from 'vitest';
 import { Upload } from '..';
-import { mockDelay, simulateFileChange, getFakeFileList, simulateDragFileChange } from '@tdesign/test/utils';
+import { mockDelay, simulateFileChange, getFakeFileList, simulateDragFileChange } from '@tdesign/internal-tests/utils';
 import { getUploadServer } from './request';
 
 describe('Upload Component', () => {
